+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [\#46](https://github.com/mllam/weather-model-graphs/pull/46), @leifdenby & @Joltsy10
 - Add `__version__` attribute to the package init
   [\#56](https://github.com/mllam/weather-model-graphs/pull/56) @AdMub
+- Add runtime and scaling benchmarks with CLI,
+  [\#117](https://github.com/mllam/weather-model-graphs/pull/117), @leifdenby
 
 ## [v0.3.0](https://github.com/mllam/weather-model-graphs/releases/tag/v0.3.0)